Add experimental capability to set a speed profile
This follows OpenCorsairLink's implementation for devices of the same
period.  Only the fan channel _might_ have this capability.

The profile is normalized, a safeguard is put in place for critical
liquid temperature (60°C) and, finally, it is optimized to use as many
of the available (6) points as possible.

Using the pseudo-channel 'cooler' will first set the pump to whatever
the fans will be at a reference temperature, which has initially been
chosen to be 50°C.